poetry/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
docopt==0.6.2
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pipreqs==0.4.11
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==2.1.1
yarg==0.1.9
zipp==3.8.0
